fix: error while installing opencv (#241)
* refactor: remove unnecessary 'sudo' from OpenCV installation command

* feat: run build

* feat: Update OpenCV build command to use sccache

The previous commit added support for using sccache in the OpenCV build process. However, the command to set the CMAKE_C_COMPILER_LAUNCHER and CMAKE_CXX_COMPILER_LAUNCHER variables was not correctly using the sccache path. This commit fixes that issue by replacing the hardcoded path with the ${SCCACHE_PATH} variable.

* fix: Use sccache path variable in OpenCV build command